pyproject.toml
[project]
name = "travelplay-app"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langchain==0.2.12",
  "langchain-openai==0.1.23",
  "openai==1.40.2",
  "httpx==0.27.2",     # pin to avoid proxies kwarg issue
  "python-dotenv==1.0.1",
  "pydantic==2.8.2",
  "tenacity==8.5.0",
  "fpdf2==2.7.9",
]

[tool.black]
line-length = 100

[tool.ruff]
lint.select = ["E","F","I"]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
